requirements.txt
openai_whisper==20230314
streamlit==1.22.0
